app/requirements/dev.txt
-r base.txt

django-debug-toolbar
factory-boy
